code/requirements.txt
cityscapesscripts==2.2.2
# cityscapesscripts[gui]==2.2.2
opencv-python-headless==4.9.0.80
# opencv-python==4.9.0.80
pillow==10.3.0
numpy==1.26.4
# tqdm==2.2.3
scipy==1.14.1 

torch==2.2.0
torchvision==0.17.0


PyQt5==5.15.10
PySide2==5.13.2

pyyaml==6.0.1